move slots & fix empty imported css files
