+++ src/App.vue
@@ -149,13 +149,13 @@
   import MySelect                             from '@components/MySelect.vue'
   import MyTab                                from '@components/MyTab.vue'
 
-  import SlotConfigurationCategories          from '@components/SlotConfigurationCategories.vue'
-  import SlotConfigurationExport              from '@components/SlotConfigurationExport.vue'
-  import SlotConfigurationFullScreen          from '@components/SlotConfigurationFullScreen.vue'
-  import SlotConfigurationImport              from '@components/SlotConfigurationImport.vue'
-  import SlotConfigurationTabsActive          from '@components/SlotConfigurationTabsActive.vue'
-  import SlotAddNewProduct                    from '@components/SlotAddNewProduct.vue'
-  import slotShoppingList                     from '@components/SlotShoppingList.vue'
+  import SlotConfigurationCategories          from '@slots/ConfigurationCategories.vue'
+  import SlotConfigurationExport              from '@slots/ConfigurationExport.vue'
+  import SlotConfigurationFullScreen          from '@slots/ConfigurationFullScreen.vue'
+  import SlotConfigurationImport              from '@slots/ConfigurationImport.vue'
+  import SlotConfigurationTabsActive          from '@slots/ConfigurationTabsActive.vue'
+  import SlotAddNewProduct                    from '@slots/AddNewProduct.vue'
+  import slotShoppingList                     from '@slots/ShoppingList.vue'
 
   import Swal                                 from 'sweetalert2'
   import { computed, ref, watch, onMounted }  from 'vue'
