+++ src/assets/css/MyImage.vue.css
@@ -1,0 +1,1 @@
+/* Archivo css de MyImage.vue*/
