+++ src/assets/css/SlotAddNewProduct.vue.css
@@ -1,0 +1,1 @@
+/* Archivo css de SlotAddNewProduct.vue */
