+++ src/assets/css/SlotConfigurationExport.vue.css
@@ -1,0 +1,1 @@
+/* Archivo css de SlotConfigurationExport.vue */
