+++ src/assets/css/SlotConfigurationFullScreen.vue.css
@@ -1,0 +1,1 @@
+/* Archivo css de SlotConfigurationFullScreen.vue */
