+++ src/assets/css/SlotConfigurationImport.vue.css
@@ -1,0 +1,1 @@
+/* Archivo css de SlotConfigurationImport.vue */
