+++ src/assets/css/SlotConfigurationTabsActive.vue.css
@@ -1,0 +1,1 @@
+/* Archivo css de SlotConfigurationTabsActive.vue */
