+++ src/assets/css/SlotShoppingList.vue.css
@@ -1,0 +1,1 @@
+/* Archivo css de SlotShoppingList.vue */
