+++ src/slots/AddNewProduct.vue
