+++ src/slots/ConfigurationCategories.vue
