+++ src/slots/ConfigurationFullScreen.vue
