+++ src/slots/ConfigurationTabsActive.vue
