+++ src/slots/ShoppingList.vue
