+++ vite.config.mjs
@@ -27,6 +27,7 @@ export default defineConfig({
       '@assets':      path.resolve(__dirname, 'src/assets'),
       '@css':         path.resolve(__dirname, 'src/assets/css'),
       '@components':  path.resolve(__dirname, 'src/components'),
+      '@slots':       path.resolve(__dirname, 'src/slots'), 
     },
   },
 });
